Added UserService.DeleteUser and started changing UseCases, BaseRequest not needed now, because exist ICurrentUserService, BaseResponse don`t needed now, because exist Result

+++ FinanceManager/FinanceManager.Domain/Services/Finances/FinanceService.cs
@@ -180,6 +180,9 @@ public class FinanceService : BaseService, IFinanceService
 
     public async Task DeleteFinanceOperationAsync(Guid id)
     {
+        if(id == Guid.Empty)
+            throw new ArgumentNullException(nameof(id));
+
         _financeOperationRepository.Delete(id);
         await _unitOfWork.SaveChangesAsync();
     }
